+++ xoutil/context.py
@@ -2,7 +2,7 @@
 #----------------------------------------------------------------------
 # xoutil.context
 #----------------------------------------------------------------------
-# Copyright (c) 2011, 2012 Merchise Autrement
+# Copyright (c) 2011, 2012 Medardo Rodríguez
 # All rights reserved.
 #
 # Author: Medardo Rodriguez
@@ -15,9 +15,51 @@
 # Created on Mar 9, 2011
 #
 
+
 '''
 A context manager for execution context flags.
+
+Use as:
+
+    >>> from xoutil.context import context
+    >>> with context('somename'):
+    ...     if context['somename']:
+    ...         print('In context somename')
+    In context somename
+
+Note the difference creating the context and checking it.
+
+If module `zope.interface` is installed, then you may ask for
+interfaces in context::
+
+   >>> from zope.interface import Interface, implementer
+   >>> class IFoo(Interface):
+   ...    pass
+
+   >>> class IBar(IFoo):
+   ...    pass
+
+   >>> @implementer(IBar)
+   ... class Bar(object):
+   ...    pass
+
+   >>> bar = Bar()
+   >>> ham = Bar()
+
+   >>> with context(bar):
+   ...    if context[IFoo]:
+   ...        print('IFoo')
+   IFoo
+
+Notice that the context object is *not* the name::
+
+    >>> with context(bar) as ctx:
+    ...    if bar is not ctx:
+    ...        print('bar not ctx')
+    bar not ctx
+
 '''
+
 
 from __future__ import (division as _py3_division,
                         print_function as _py3_print,
@@ -25,12 +67,20 @@ from __future__ import (division as _py3_division,
 
 from threading import local
 
+from xoutil.collections import OrderedDict
+from xoutil.compat import iteritems_
+
+try:
+    from zope.interface import Interface
+except ImportError:
+    Interface = None
+
 
 
 class LocalData(local):
     def __init__(self):
         super(LocalData, self).__init__()
-        self.contexts = {}
+        self.contexts = OrderedDict()
 
 _data = LocalData()
 
@@ -38,7 +88,21 @@ _data = LocalData()
 
 class MetaContext(type):
     def __getitem__(self, name):
-        return _data.contexts.get(name, _null_context)
+        result = _data.contexts.get(name, None)
+        if result:
+            return result
+        elif Interface and not result and type(name) is type(Interface):
+            candidates = list((len(type(which).mro()), context)
+                               for which, context in iteritems_(_data.contexts)
+                               if name.providedBy(which))
+            if candidates:
+                # Returns the most specific and last
+                candidates.sort(key=lambda (depth, cls): depth, reverse=True)
+                return candidates[0][-1]
+            else:
+                return _null_context
+        else:
+            return _null_context
 
 
     def __contains__(self, name):
@@ -81,6 +145,7 @@ class Context(object):
             res.name = name
             res.data = data
             res.count = 0
+            res._events = []
         elif data:
             res.data.update(data)
         return res
@@ -97,9 +162,22 @@ class Context(object):
     def __exit__(self, exc_type, exc_value, traceback):
         self.count -= 1
         if self.count == 0:
+            for event in self.events:
+                event(self)
             del _data.contexts[self.name]
         return False
 
+
+    @property
+    def events(self):
+        return self._events
+
+    @events.setter
+    def events(self, value):
+        self._events = list(set(value))
+
+
+# A simple alias for Context
 context = Context