+++ xoutil/formatter.py
@@ -3,7 +3,7 @@
 #----------------------------------------------------------------------
 # xoutil.formatter
 #----------------------------------------------------------------------
-# Copyright (c) 2009-2011 Merchise Autrement
+# Copyright (c) 2009-2011 Medardo Rodríguez
 # All rights reserved.
 #
 # Author: Medardo Rodriguez
@@ -23,9 +23,11 @@ from __future__ import (division as _py3_division,
                         absolute_import as _py3_abs_imports)
 
 
+
 class DelimiterFactory(object):
     def __new__(cls, owner, key, start, end):
         return key
+
 
 
 class BaseFactory(object):
@@ -37,9 +39,11 @@ class BaseFactory(object):
         self.start = start
 
 
+
 class MapFactory(BaseFactory):
     def __call__(self, mapping):
         return unicode(mapping[self.key])
+
 
 
 class PyFactory(BaseFactory):
@@ -48,8 +52,10 @@ class PyFactory(BaseFactory):
                                         compile(key, '', 'eval'),
                                         start, end)
 
+
     def __call__(self, mapping):
         return unicode(eval(self.key, mapping))
+
 
 
 class InvalidFactory(object):
@@ -71,6 +77,7 @@ class InvalidFactory(object):
         raise ValueError('Invalid place-holder in string: line "%d", col "%d"' % (line, col))
 
 
+
 class _TemplateClass(type):
     'Metaclass for Template.'
 
@@ -82,6 +89,7 @@ class _TemplateClass(type):
                ('python', r'[^{}]+', '{\?%s}', PyFactory),
                ('invalid', r'.', '%s', InvalidFactory))
 
+
     def __init__(cls, name, bases, attrs):
         import re
         super(_TemplateClass, cls).__init__(name, bases, attrs)
@@ -92,6 +100,7 @@ class _TemplateClass(type):
         rexp = r'(?P<delimiter>%s)(?:%s)' % (re.escape(cls.delimiter), b'|'.join(alters))
         cls.pattern = re.compile(rexp, re.IGNORECASE | re.VERBOSE)
         cls.factories = factories
+
 
 
 class Template(object):
@@ -108,7 +117,18 @@ class Template(object):
     If you need repetition or other flow-control sentences you should use
     other templating system.
 
-    TODO: [med] Properly document with examples.
+    If you enclose and expression within ``${?...}`` it will be evaluated as a
+    python expression. Simple variables are allowed just with ``$var`` or
+    ``${var}``::
+
+        >>> tpl = Template(u'${?1 + 1} is 2, and ${?x + x} is $x + ${x}')
+        >>> tpl % dict(x=4)
+        u'2 is 2, and 8 is 4 + 4'
+
+    The mapping may be given by calling the template::
+
+        >>> tpl(x=5)
+        u'2 is 2, and 10 is 5 + 5'
     '''
 
     __metaclass__ = _TemplateClass
@@ -134,8 +154,10 @@ class Template(object):
                     self._append(aux)
                 valid = False
 
+
     def __str__(self):
         return '%s(%s)' % (self.__class__.__name__, self.template)
+
 
     def __call__(self, mapping={}, **kwargs):
         # TODO: Don't update if object
@@ -148,12 +170,15 @@ class Template(object):
                 res += item(kwargs)
         return res
 
+
     def __mod__(self, mapping):
         '''template % {'x':1}'''
         return self(mapping)
 
+
     def substitute(self, mapping={}, **kwargs):
         return self(mapping, **kwargs)
+
 
     def safe_substitute(self, mapping={}, **kwargs):
         # TODO: Don't update if object
@@ -172,11 +197,13 @@ class Template(object):
                         res += ''    # item.match
         return res
 
+
     def _append(self, item):
         if isinstance(item, basestring) and self.items and isinstance(self.items[-1], basestring):
             self.items[-1] += item
         else:
             self.items.append(item)
+
 
     def _GetFactory(self, token):
         keys = self.factories.keys()
@@ -190,6 +217,7 @@ class Template(object):
             else:
                 i += 1
         return res, aux
+
 
 
 def count(source, chars):