+++ xoutil/objects.py
@@ -3,7 +3,7 @@
 #----------------------------------------------------------------------
 # xoutil.objutil
 #----------------------------------------------------------------------
-# Copyright (c) 2012 Merchise Autrement
+# Copyright (c) 2012 Medardo Rodríguez
 # All rights reserved.
 #
 # Author: Medardo Rodríguez