+++ xoutil/release.py
@@ -3,7 +3,7 @@
 #----------------------------------------------------------------------
 # xoutil.release
 #----------------------------------------------------------------------
-# Copyright (c) 2012 Merchise Autrement
+# Copyright (c) 2012 Medardo Rodríguez
 # All rights reserved.
 #
 # This is free software; you can redistribute it and/or modify it under the
@@ -18,4 +18,4 @@ from __future__ import (division as _py3_division,
                         unicode_literals as _py3_unicode,
                         absolute_import as _py3_abs_imports)
 
-VERSION = '1.1.1'
+VERSION = '1.1.4'