+++ xoutil/tests/test_decorators.py
@@ -3,7 +3,7 @@
 #----------------------------------------------------------------------
 # xoutil.tests.test_decorators
 #----------------------------------------------------------------------
-# Copyright (c) 2011 Merchise Autrement
+# Copyright (c) 2011 Medardo Rodríguez
 # All rights reserved.
 #
 # This is free software; you can redistribute it and/or modify it under