+++ xoutil/trees/base.py
@@ -2,7 +2,7 @@
 #----------------------------------------------------------------------
 # xoutil.trees.base
 #----------------------------------------------------------------------
-# Copyright (c) 2011 Merchise Autrement
+# Copyright (c) 2011 Medardo Rodríguez
 # All rights reserved.
 #
 # Author: Medardo Rodriguez