+++ docs/source/HISTORY.rst
@@ -3,18 +3,50 @@ Changelog
 
 Reversed chronological order.
 
-
 1.1 series
 ----------
+
+2012-11-01. Release 1.1.4
+~~~~~~~~~~~~~~~~~~~~~~~~~
+
+- Introduces :func:`xoutil.compat.iteritems_`,
+  :func:`xoutil.compat.iterkeys_`
+  and :func:`xoutil.compat.itervalues_`.
+
+- :mod:`execution context <xoutil.context>` are now aware of
+  `zope.interface` interfaces; so that you may ask for a context name
+  implementing a given interface, instead of the name itself.
+
+- Improves xoutil.formatter documentation.
+
+- Several fixes to xoutil.aop.classical. It has sudden backwards
+  incompatible changes.
+
+- `before` and `after` methods may use the `*args, **kwargs` idiom
+  to get the passed arguments of the weaved method.
+
+- Several minor fixes: Invalid warning about Unset not in xoutil.types
+
+2012-08-22. Release 1.1.3
+~~~~~~~~~~~~~~~~~~~~~~~~~
+
+- Adds function :func:`xoutil.fs.rmdirs` that removes empty dirs.
+
+- Adds functions :func:`xoutil.string.safe_join`,
+  :func:`xoutil.string.safe_encode`, :func:`xoutil.string.safe_decode`,
+  and :func:`xoutil.string.safe_strip`; and the class
+  :class:`xoutil.string.SafeFormatter`.
+
+- Adds function :func:`xoutil.cpystack.iter_frames`.
 
 2012-07-11. Release 1.1.2
 ~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-- Fixes all copyrights notices and chooses MIT as the license model for xoutil 
-  releases.
-  
+- Fixes all copyrights notices and chooses the PSF License for Python 3.2.3
+  as the license model for xoutil releases.
+
 - All releases from now on will be publicly available at github_.
-  
+
 .. _github: https://github.com/merchise-autrement/xoutil/
 
 2012-07-06. Release 1.1.1
@@ -46,7 +78,6 @@ Reversed chronological order.
 
 1.0 series
 ----------
-
 
 2012-06-15. Release 1.0.30
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~